Extra trim options

+++ common/src/main/java/org/popcraft/chunky/command/TrimCommand.java
@@ -2,6 +2,8 @@ package org.popcraft.chunky.command;
 
 import org.popcraft.chunky.Chunky;
 import org.popcraft.chunky.Selection;
+import org.popcraft.chunky.nbt.CompoundTag;
+import org.popcraft.chunky.nbt.LongTag;
 import org.popcraft.chunky.platform.Sender;
 import org.popcraft.chunky.platform.World;
 import org.popcraft.chunky.shape.Shape;
@@ -11,6 +13,7 @@ import org.popcraft.chunky.util.ChunkCoordinate;
 import org.popcraft.chunky.util.Formatting;
 import org.popcraft.chunky.util.Input;
 import org.popcraft.chunky.util.TranslationKey;
+import org.popcraft.chunky.world.RegionFile;
 
 import java.io.IOException;
 import java.io.RandomAccessFile;
@@ -79,6 +82,9 @@ public class TrimCommand implements ChunkyCommand {
                 return;
             }
         }
+        final boolean inside = arguments.next().map(String::toLowerCase).map("inside"::equals).orElse(false);
+        final int inhabitedTime = arguments.next().flatMap(Input::tryIntegerSuffixed).orElse(Integer.MAX_VALUE);
+        final boolean inhabitedTimeCheck = inhabitedTime < Integer.MAX_VALUE;
         final Selection selection = chunky.getSelection().build();
         final Shape shape = ShapeFactory.getShape(selection);
         final Runnable deletionAction = () -> chunky.getScheduler().runTask(() -> {
@@ -91,17 +97,17 @@ public class TrimCommand implements ChunkyCommand {
             try {
                 if (regionPath.isPresent()) {
                     try (final Stream<Path> regionWalker = Files.walk(regionPath.get())) {
-                        regionWalker.forEach(region -> deleted.getAndAdd(checkRegion(region, shape)));
+                        regionWalker.forEach(region -> deleted.getAndAdd(checkRegion(region, shape, inside, inhabitedTimeCheck, inhabitedTime)));
                     }
                 }
                 if (poiPath.isPresent()) {
                     try (final Stream<Path> poiWalker = Files.walk(poiPath.get())) {
-                        poiWalker.forEach(region -> checkRegion(region, shape));
+                        poiWalker.forEach(region -> checkRegion(region, shape, inside, inhabitedTimeCheck, inhabitedTime));
                     }
                 }
                 if (entitiesPath.isPresent()) {
                     try (final Stream<Path> entityWalker = Files.walk(entitiesPath.get())) {
-                        entityWalker.forEach(region -> checkRegion(region, shape));
+                        entityWalker.forEach(region -> checkRegion(region, shape, inside, inhabitedTimeCheck, inhabitedTime));
                     }
                 }
             } catch (IOException e) {
@@ -111,20 +117,20 @@ public class TrimCommand implements ChunkyCommand {
             sender.sendMessagePrefixed(TranslationKey.TASK_TRIM, deleted.get(), selection.world().getName(), String.format("%.3f", totalTime / 1e3f));
         });
         chunky.setPendingAction(sender, deletionAction);
-        sender.sendMessagePrefixed(TranslationKey.FORMAT_TRIM_CONFIRM, selection.world().getName(), translate("shape_" + selection.shape()), Formatting.number(selection.centerX()), Formatting.number(selection.centerZ()), Formatting.radius(selection), "/chunky confirm");
+        sender.sendMessagePrefixed(inside ? TranslationKey.FORMAT_TRIM_CONFIRM_INSIDE : TranslationKey.FORMAT_TRIM_CONFIRM, selection.world().getName(), translate("shape_" + selection.shape()), Formatting.number(selection.centerX()), Formatting.number(selection.centerZ()), Formatting.radius(selection), "/chunky confirm");
     }
 
-    private int checkRegion(final Path region, final Shape shape) {
+    private int checkRegion(final Path region, final Shape shape, final boolean inside, final boolean inhabitedTimeCheck, final int inhabitedTime) {
         final Optional<ChunkCoordinate> regionCoordinate = tryRegionCoordinate(region);
         if (regionCoordinate.isEmpty()) {
             return 0;
         }
         final int chunkX = regionCoordinate.get().x() << 5;
         final int chunkZ = regionCoordinate.get().z() << 5;
-        if (shouldDeleteRegion(shape, chunkX, chunkZ)) {
+        if (!inhabitedTimeCheck && shouldDeleteRegion(shape, inside, chunkX, chunkZ)) {
             return deleteRegion(region);
         } else {
-            return trimRegion(region, shape, chunkX, chunkZ);
+            return trimRegion(region, shape, inside, chunkX, chunkZ, inhabitedTimeCheck, inhabitedTime);
         }
     }
 
@@ -147,12 +153,12 @@ public class TrimCommand implements ChunkyCommand {
         return Optional.empty();
     }
 
-    private boolean shouldDeleteRegion(final Shape shape, final int chunkX, final int chunkZ) {
+    private boolean shouldDeleteRegion(final Shape shape, final boolean inside, final int chunkX, final int chunkZ) {
         for (int offsetX = 0; offsetX < 32; ++offsetX) {
             for (int offsetZ = 0; offsetZ < 32; ++offsetZ) {
                 final int chunkCenterX = ((chunkX + offsetX) << 4) + 8;
                 final int chunkCenterZ = ((chunkZ + offsetZ) << 4) + 8;
-                if (shape.isBounding(chunkCenterX, chunkCenterZ)) {
+                if (inside != shape.isBounding(chunkCenterX, chunkCenterZ)) {
                     return false;
                 }
             }
@@ -170,17 +176,41 @@ public class TrimCommand implements ChunkyCommand {
         }
     }
 
-    private int trimRegion(final Path region, final Shape shape, final int chunkX, final int chunkZ) {
+    private int trimRegion(final Path region, final Shape shape, final boolean inside, final int chunkX, final int chunkZ, final boolean inhabitedTimeCheck, final int inhabitedTime) {
+        int marked = 0;
         int deleted = 0;
-        try (RandomAccessFile regionFile = new RandomAccessFile(region.toFile(), "rw")) {
+        final RegionFile regionData = inhabitedTimeCheck ? new RegionFile(region.toFile()) : null;
+        try (final RandomAccessFile regionFile = new RandomAccessFile(region.toFile(), "rw")) {
             if (regionFile.length() < 4096) {
                 return 0;
             }
             for (int offsetX = 0; offsetX < 32; ++offsetX) {
                 for (int offsetZ = 0; offsetZ < 32; ++offsetZ) {
-                    final int chunkCenterX = ((chunkX + offsetX) << 4) + 8;
-                    final int chunkCenterZ = ((chunkZ + offsetZ) << 4) + 8;
-                    if (!shape.isBounding(chunkCenterX, chunkCenterZ)) {
+                    final int offsetChunkX = chunkX + offsetX;
+                    final int offsetChunkZ = chunkZ + offsetZ;
+                    final int chunkCenterX = (offsetChunkX << 4) + 8;
+                    final int chunkCenterZ = (offsetChunkZ << 4) + 8;
+                    final boolean trimChunk;
+                    if (inside) {
+                        trimChunk = shape.isBounding(chunkCenterX, chunkCenterZ);
+                    } else {
+                        trimChunk = !shape.isBounding(chunkCenterX, chunkCenterZ);
+                    }
+                    final boolean trimInhabited = regionData == null || regionData.getChunk(offsetChunkX, offsetChunkZ)
+                            .map(chunk -> {
+                                final CompoundTag compoundTag = chunk.getData();
+                                if (compoundTag == null) {
+                                    return true;
+                                }
+                                final LongTag inhabited = compoundTag.getLong("InhabitedTime").orElse(null);
+                                if (inhabited == null) {
+                                    return true;
+                                }
+                                return inhabited.value() <= inhabitedTime;
+                            })
+                            .orElse(true);
+                    if (trimChunk && trimInhabited) {
+                        ++marked;
                         final int chunkLocation = ((offsetX % 32) + (offsetZ % 32) * 32) * 4;
                         regionFile.seek(chunkLocation);
                         if (regionFile.readInt() != 0) {
@@ -193,6 +223,9 @@ public class TrimCommand implements ChunkyCommand {
             }
         } catch (IOException e) {
             e.printStackTrace();
+        }
+        if (inhabitedTimeCheck && marked == 1024) {
+            deleteRegion(region);
         }
         return deleted;
     }
